--- multimodal/src/autogluon/multimodal/models/conv_lora/adaptation_layers.py
@@ -1,133 +1,0 @@
-import math
-from typing import Optional
-
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-
-from ..adaptation_layers import LoRALayer
-from .moe_conv import MoEConv, SparseDispatcher
-
-
-class ConvLoRALinear(nn.Linear, LoRALayer):
-    """
-    Conv-LoRA incorporated in Linear Layer. Weights of linear layer are set to be frozen per default.
-
-    Parameters
-    ----------
-    in_features
-        input dimension, set to the original linear layer input dimension LoRA is replacing.
-    out_features
-        output dimension, set to the original linear layer output dimension LoRA is replacing.
-    r
-        rank r of the low-rank decomposition.
-    lora_alpha
-        Scaling factor. Can be simply set to same value as r as initialization is scaled already.
-    lora_dropout
-        Dropout probability.
-    fan_in_fan_out
-        Set this to True if the layer to replace stores weight like (fan_in, fan_out).
-    merge_weights
-        Merging weights during inference to reduce latency.
-    conv_lora_expert_num
-        The number of experts in MoE-Conv.
-
-    References
-    ----------
-    1. Zihan Zhong, Zhiqiang Tang, Tong He, Haoyang Fang, Chun Yuan,
-    "Convolution Meets LoRA: Parameter Efficient Finetuning for Segment Anything Model", 2024
-    https://arxiv.org/abs/2401.17868
-    """
-
-    def __init__(
-        self,
-        in_features: int,
-        out_features: int,
-        r: int = 0,
-        lora_alpha: int = 1,
-        lora_dropout: float = 0.0,
-        fan_in_fan_out: bool = False,  # Set this to True if the layer to replace stores weight like (fan_in, fan_out)
-        merge_weights: bool = False,
-        conv_lora_expert_num: Optional[int] = None,
-        **kwargs,
-    ):
-        nn.Linear.__init__(self, in_features, out_features, **kwargs)
-        LoRALayer.__init__(self, r=r, lora_alpha=lora_alpha, lora_dropout=lora_dropout, merge_weights=merge_weights)
-
-        self.fan_in_fan_out = fan_in_fan_out
-        # Actual trainable parameters
-        if r > 0:
-            self.lora_A = nn.Parameter(self.weight.new_zeros((r, in_features)))
-            self.lora_B = nn.Parameter(self.weight.new_zeros((out_features, r)))
-            self.scaling = self.lora_alpha / self.r
-            # Freezing the pre-trained weight matrix
-            self.weight.requires_grad = False
-
-            # MoE-Conv
-            topk = 1
-            self.lora_moe_gating = MoEConv(M=conv_lora_expert_num, d=self.r, K=topk)
-            self.lora_moe_experts = nn.ModuleList([])
-            self.upsample_ratios = list(range(1, conv_lora_expert_num + 1))
-            for upsample_ratio in self.upsample_ratios:
-                expert = nn.Conv2d(in_channels=r, out_channels=r, kernel_size=3, stride=1, padding=1, bias=True)
-                expert.bias.data.zero_()
-                self.lora_moe_experts.append(nn.Sequential(expert, nn.GELU()))
-            self.num_experts = conv_lora_expert_num
-            self.multiply_by_gates = False
-
-        self.reset_parameters()
-        if fan_in_fan_out:
-            self.weight.data = self.weight.data.T
-
-    def reset_parameters(self):
-        nn.Linear.reset_parameters(self)
-        if hasattr(self, "lora_A"):
-            # initialize A the same way as the default for nn.Linear and B to zero
-            nn.init.kaiming_uniform_(self.lora_A, a=math.sqrt(5))
-            nn.init.zeros_(self.lora_B)
-
-    def T(self, w):
-        return w.T if self.fan_in_fan_out else w
-
-    def forward(self, x: torch.Tensor):
-        result = F.linear(x, self.T(self.weight), bias=self.bias)
-        if self.r > 0:
-            lora_res = self.lora_dropout(x) @ self.lora_A.T
-            dim = lora_res.dim()
-            if dim == 3:
-                B, L, C = lora_res.size()
-                H = W = int(math.sqrt(L))
-                lora_res = lora_res.reshape(B, H, W, C)
-            else:
-                H, W = lora_res.size()[1:3]
-
-            # Calculate the gating values.
-            lora_res = lora_res.permute(0, 3, 1, 2).contiguous()
-            gates, moe_loss = self.lora_moe_gating(lora_res)
-
-            # Distribute data samples to experts.
-            dispatcher = SparseDispatcher(self.num_experts, gates)
-            expert_inputs = dispatcher.dispatch(lora_res)
-            expert_outputs = []
-            for i in range(self.num_experts):
-                if len(expert_inputs[i]) == 0:
-                    continue
-                upsample_ratio = self.upsample_ratios[i]
-                cur_res = expert_inputs[i]
-                if upsample_ratio != 1:
-                    cur_res = F.interpolate(cur_res, scale_factor=upsample_ratio, mode="bicubic")
-                cur_res = self.lora_moe_experts[i](cur_res)
-                if upsample_ratio != 1:
-                    cur_res = F.interpolate(cur_res, size=(int(H), int(W)), mode="bicubic")
-                expert_outputs.append(cur_res)
-
-            # Combine data samples after processing by each expert.
-            temp_lora_res = dispatcher.combine(expert_outputs, multiply_by_gates=self.multiply_by_gates)
-            lora_res = lora_res + temp_lora_res
-
-            lora_res = lora_res.permute(0, 2, 3, 1).contiguous()
-            if dim == 3:
-                lora_res = lora_res.reshape(B, L, C)
-            result += (lora_res @ self.lora_B.T) * self.scaling
-
-        return result, moe_loss
