--- multimodal/src/autogluon/multimodal/models/conv_lora/moe_conv.py
@@ -1,238 +1,0 @@
-from typing import Any, Callable, Dict, Iterator, List, Mapping, Optional, Set, Tuple, TypeVar, Union, overload
-
-import numpy as np
-import torch
-from torch import nn
-from torch.distributions.normal import Normal
-
-
-class MoEConv(nn.Module):
-    def __init__(self, d, M=4, K=1, noisy_gating=True):
-        """Constructor
-        Args:
-            d: input channel dimensionality.
-            M: the number of experts.
-            K: the number of chosen experts for each forward pass.
-        """
-        super(MoEConv, self).__init__()
-        self.M = M
-        self.k = K
-        self.gap = nn.AdaptiveAvgPool2d((1, 1))  # global average pooling
-
-        self.noisy_gating = noisy_gating
-
-        self.w_gate = nn.Parameter(torch.zeros(d, M), requires_grad=True)
-        self.w_noise = nn.Parameter(torch.zeros(d, M), requires_grad=True)
-
-        self.softplus = nn.Softplus()
-        self.softmax = nn.Softmax(1)
-        self.register_buffer("mean", torch.tensor([0.0]))
-        self.register_buffer("std", torch.tensor([1.0]))
-        assert self.k <= self.M
-
-    def forward(self, feats, loss_coef=1e-2, noise_epsilon=1e-2):
-        batch_size = feats.shape[0]
-
-        feats_S = self.gap(feats).view(batch_size, -1)
-
-        clean_logits = feats_S @ self.w_gate
-        if self.noisy_gating and self.training:
-            raw_noise_stddev = feats_S @ self.w_noise
-            noise_stddev = self.softplus(raw_noise_stddev) + noise_epsilon
-            noisy_logits = clean_logits + (torch.randn_like(clean_logits) * noise_stddev)
-            logits = noisy_logits
-        else:
-            logits = clean_logits
-
-        top_logits, top_indices = logits.topk(min(self.k + 1, self.M), dim=1)
-        top_k_logits = top_logits[:, : self.k]
-        top_k_indices = top_indices[:, : self.k]
-        top_k_gates = self.softmax(top_k_logits)
-        zeros = torch.zeros_like(logits, requires_grad=True).float()
-        gates = zeros.scatter(1, top_k_indices, top_k_gates).to(logits.dtype)
-
-        if self.noisy_gating and self.k < self.M and self.training:
-            load = (self._prob_in_top_k(clean_logits, noisy_logits, noise_stddev, top_logits)).sum(0)
-        else:
-            load = self._gates_to_load(gates)
-
-        importance = gates.sum(0)
-        loss = self.cv_squared(importance) + self.cv_squared(load)
-        loss *= loss_coef
-
-        return gates, loss
-
-    def _gates_to_load(self, gates):
-        """Compute the true load per expert, given the gates.
-        The load is the number of examples for which the corresponding gate is >0.
-        Args:
-        gates: a `Tensor` of shape [batch_size, n]
-        Returns:
-        a float32 `Tensor` of shape [n]
-        """
-        return (gates > 0).sum(0)
-
-    def cv_squared(self, x):
-        """The squared coefficient of variation of a sample.
-        Useful as a loss to encourage a positive distribution to be more uniform.
-        Epsilons added for numerical stability.
-        Returns 0 for an empty Tensor.
-        Args:
-        x: a `Tensor`.
-        Returns:
-        a `Scalar`.
-        """
-        eps = 1e-10
-
-        if x.shape[0] == 1:
-            return torch.tensor([0], device=x.device, dtype=x.dtype)
-        return x.float().var() / (x.float().mean() ** 2 + eps)
-
-    def _prob_in_top_k(self, clean_values, noisy_values, noise_stddev, noisy_top_values):
-        """Helper function to NoisyTopKGating.
-        Computes the probability that value is in top k, given different random noise.
-        This gives us a way of backpropagating from a loss that balances the number
-        of times each expert is in the top k experts per example.
-        In the case of no noise, pass in None for noise_stddev, and the result will
-        not be differentiable.
-        Args:
-        clean_values: a `Tensor` of shape [batch, n].
-        noisy_values: a `Tensor` of shape [batch, n].  Equal to clean values plus
-          normally distributed noise with standard deviation noise_stddev.
-        noise_stddev: a `Tensor` of shape [batch, n], or None
-        noisy_top_values: a `Tensor` of shape [batch, m].
-           "values" Output of tf.top_k(noisy_top_values, m).  m >= k+1
-        Returns:
-        a `Tensor` of shape [batch, n].
-        """
-        batch = clean_values.size(0)
-        m = noisy_top_values.size(1)
-        top_values_flat = noisy_top_values.flatten()
-
-        threshold_positions_if_in = torch.arange(batch, device=clean_values.device) * m + self.k
-        threshold_if_in = torch.unsqueeze(torch.gather(top_values_flat, 0, threshold_positions_if_in), 1)
-        is_in = torch.gt(noisy_values, threshold_if_in)
-        threshold_positions_if_out = threshold_positions_if_in - 1
-        threshold_if_out = torch.unsqueeze(torch.gather(top_values_flat, 0, threshold_positions_if_out), 1)
-        # is each value currently in the top k.
-        normal = Normal(self.mean, self.std)
-        prob_if_in = normal.cdf((clean_values - threshold_if_in) / noise_stddev)
-        prob_if_out = normal.cdf((clean_values - threshold_if_out) / noise_stddev)
-        prob = torch.where(is_in, prob_if_in, prob_if_out)
-        return prob
-
-
-class SparseDispatcher(object):
-    """Helper for implementing a mixture of experts.
-    The purpose of this class is to create input minibatches for the
-    experts and to combine the results of the experts to form a unified
-    output tensor.
-    There are two functions:
-    dispatch - take an input Tensor and create input Tensors for each expert.
-    combine - take output Tensors from each expert and form a combined output
-      Tensor.  Outputs from different experts for the same batch element are
-      summed together, weighted by the provided "gates".
-    The class is initialized with a "gates" Tensor, which specifies which
-    batch elements go to which experts, and the weights to use when combining
-    the outputs.  Batch element b is sent to expert e iff gates[b, e] != 0.
-    The inputs and outputs are all two-dimensional [batch, depth].
-    Caller is responsible for collapsing additional dimensions prior to
-    calling this class and reshaping the output to the original shape.
-    See common_layers.reshape_like().
-    Example use:
-    gates: a float32 `Tensor` with shape `[batch_size, num_experts]`
-    inputs: a float32 `Tensor` with shape `[batch_size, input_size]`
-    experts: a list of length `num_experts` containing sub-networks.
-    dispatcher = SparseDispatcher(num_experts, gates)
-    expert_inputs = dispatcher.dispatch(inputs)
-    expert_outputs = [experts[i](expert_inputs[i]) for i in range(num_experts)]
-    outputs = dispatcher.combine(expert_outputs)
-    The preceding code sets the output for a particular example b to:
-    output[b] = Sum_i(gates[b, i] * experts[i](inputs[b]))
-    This class takes advantage of sparsity in the gate matrix by including in the
-    `Tensor`s for expert i only the batch elements for which `gates[b, i] > 0`.
-
-    References
-    ----------
-    1. Noam Shazeer, Azalia Mirhoseini, Krzysztof Maziarz, Andy Davis, Quoc Le, Geoffrey Hinton, Jeff Dean,
-    "Outrageously Large Neural Networks: The Sparsely-Gated Mixture-of-Experts Layer", 2017
-    https://arxiv.org/abs/1701.06538
-    2. Code: https://github.com/davidmrau/mixture-of-experts/blob/master/moe.py
-    """
-
-    def __init__(self, num_experts, gates):
-        """Create a SparseDispatcher."""
-        self._gates = gates
-        self._num_experts = num_experts
-        # sort experts
-        sorted_experts, index_sorted_experts = torch.nonzero(gates).sort(0)
-        # drop indices
-        _, self._expert_index = sorted_experts.split(1, dim=1)
-        # get according batch index for each expert
-        self._batch_index = torch.nonzero(gates)[index_sorted_experts[:, 1], 0]
-        # calculate num samples that each expert gets
-        self._part_sizes = (gates > 0).sum(0).tolist()
-        # expand gates to match with self._batch_index
-        gates_exp = gates[self._batch_index.flatten()]
-        self._nonzero_gates = torch.gather(gates_exp, 1, self._expert_index)
-
-    def dispatch(self, inp):
-        """Create one input Tensor for each expert.
-        The `Tensor` for a expert `i` contains the slices of `inp` corresponding
-        to the batch elements `b` where `gates[b, i] > 0`.
-        Args:
-          inp: a `Tensor` of shape "[batch_size, <extra_input_dims>]`
-        Returns:
-          a list of `num_experts` `Tensor`s with shapes
-            `[expert_batch_size_i, <extra_input_dims>]`.
-        """
-
-        # assigns samples to experts whose gate is nonzero
-
-        # expand according to batch index so we can just split by _part_sizes
-        inp_exp = inp[self._batch_index].squeeze(1)  # [bs * num_of chosen experts, dim]
-        return torch.split(inp_exp, self._part_sizes, dim=0)
-
-    def combine(self, expert_out, multiply_by_gates=True):
-        """Sum together the expert output, weighted by the gates.
-        The slice corresponding to a particular batch element `b` is computed
-        as the sum over all experts `i` of the expert output, weighted by the
-        corresponding gate values.  If `multiply_by_gates` is set to False, the
-        gate values are ignored.
-        Args:
-          expert_out: a list of `num_experts` `Tensor`s, each with shape
-            `[expert_batch_size_i, <extra_output_dims>]`.
-          multiply_by_gates: a boolean
-        Returns:
-          a `Tensor` with shape `[batch_size, <extra_output_dims>]`.
-        """
-        # apply exp to expert outputs, so we are not longer in log space
-        stitched = torch.cat(expert_out, 0).exp()
-
-        if multiply_by_gates:
-            # stitched = stitched.mul(self._nonzero_gates)
-            stitched = stitched.mul(self._nonzero_gates.unsqueeze(-1).unsqueeze(-1))
-        # zeros = torch.zeros(self._gates.size(0), expert_out[-1].size(1), requires_grad=True, device=stitched.device)
-        zeros = torch.zeros(
-            self._gates.size(0),
-            expert_out[-1].size()[1],
-            expert_out[-1].size()[2],
-            expert_out[-1].size()[3],
-            requires_grad=True,
-            device=stitched.device,
-        )
-        # combine samples that have been processed by the same k experts
-        combined = zeros.index_add(0, self._batch_index, stitched.float())
-        # add eps to all zero values in order to avoid nans when going back to log space
-        combined[combined == 0] = np.finfo(float).eps
-        # back to log space
-        return combined.log()
-
-    def expert_to_gates(self):
-        """Gate values corresponding to the examples in the per-expert `Tensor`s.
-        Returns:
-          a list of `num_experts` one-dimensional `Tensor`s with type `tf.float32`
-              and shapes `[expert_batch_size_i]`
-        """
-        # split nonzero gates for each expert
-        return torch.split(self._nonzero_gates, self._part_sizes, dim=0)
